pyproject.toml
[project]
name = "agno-learn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "agno[postgres]>=2.4.6",
    "ensembledata>=0.2.11",
    "google-genai>=0.2.0",
    "openai>=2.16.0",
    "psycopg[binary]>=3.1.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.46",
    "streamlit>=1.30.0",
]
